[refactor] 진입점 컴포넌트 추가
